fix(simulation): refract file structure
