+++ src/simulation/defense/mine.rs
@@ -1,0 +1,21 @@
+pub struct Mine {
+    pub id: i32,
+    pub mine_type: i32,
+    pub damage: i32,
+    pub radius: i32,
+    pub is_activated: i32,
+    pub x_position: i32,
+    pub y_position: i32,
+}
+
+impl Mine {
+    // #[allow(dead_code)]
+    // pub fn new() -> Self {
+    //     todo!()
+    // }
+
+    #[allow(dead_code)]
+    pub fn simulate() {
+        todo!()
+    }
+}
